fix i686/MinGW portability: cross-build now produces working exes
- pic_stdint.h/pic_config.h: gate MSVC-era stdint + lround workarounds to
  actual old MSVC (MinGW has the real headers/functions)
- pic_usb_linux.cpp: LIBUSB_CALL on transfer callbacks (stdcall on Windows,
  no-op elsewhere), pic_microsleep instead of POSIX sleep
- ef_pico.cpp: widen strip delta before abs() -- 32-bit long made the
  long-minus-unsigned diff wrap (and std::abs ambiguous)
- toolchain: -msse2 (win32 thread code uses SSE denormal intrinsics)

exe imports: kernel32/shell32/ws2_32/msvcrt + pico_decoder_1_0_0.dll only.
Linux build + tests still green.

+++ eigenapi/src/ef_pico.cpp
@@ -264,7 +264,10 @@ void EF_Pico::Delegate::kbd_strip(unsigned long long t, unsigned s) {
             // s_count_ = 80;
 
             // active, until we go below threshold
-            if (s < s_threshold_ || std::abs(long(s) - s_last_) > 200) {
+            // widen both sides before subtracting: with 32-bit long the
+            // long-minus-unsigned difference converts back to unsigned and
+            // wraps (and std::abs(unsigned long) is ambiguous)
+            if (s < s_threshold_ || std::abs(static_cast<long long>(s) - static_cast<long long>(s_last_)) > 200) {
                 // possibly ending...
                 s_state_ = 3;
                 s_count_ = 80;
